move some common functions to Utils/ for clarity
